pythonService/requirements.txt
flask
gunicorn
kafka-python